Homologação
Cadastrar
	•	Retirar a logo PayPal da câmera do cartão
	•	Trocar título para cadastro de login (100%)
	•	mensagem de erro na senha (100%)
	•	Link para termos de uso e privacidade (100%)
	•	Na validação dos campos não esta sendo dinâmica  (100%)

Lista de Cartões
	•	Salvar imagens em cache

Ajustes de Segurança
	•	Padronizar as açoēs (100%)
	•	Notificação de bloqueio do cartão (100%)

Detalhe do Produto
	•	Alterar ordem entre valores e descrição
	•	Aumentar numero de linhas do descrição do  Produto
	•	Logo do parceiro

Carrinho
	•	Detalhe do produto
	•	Botão esvaziar carrinho

Detalhe do Pedido
	•	Logo do parceiro

Menu Flutuante

  1- Geral
	•	Trocar verificação  Id produto por id plataforma (100%)
	•	Label componente animação (100%)

 2- Trocar Senha
	•	Implementar Serviço (100%)
	•	Trocar Mensagem Perda ou Roubo: (100%)
“Deseja realmente informara perda ou Roubo dece cartão ?
Este cartão será cancelado e um novo cartão será emitido.”
	•	Voltar para a tela de detalhe do cartão. (100%)
	•	Transferencia de cartão de crédito, ao chamar dados do cliente iniciar loading. (100%)
	•	Transferir para outro cartão colocar (R$) tarifa (100%)
	•	Hash do numero do cartão (100%)
	•	Senha do Cartão criptografada (100%)
	•	Confirmação de Transferencia(100%)
	•	Remoção do nome do transferir cartões(100%)

 3- Transferencia CC
	•	Mascara conta aumentar quantidade de números (100%)
	•	Apresentar CPF do usuario (100%)
	•	Tarifa adicionar (R$) (100%)
	•	Validar preenchimento dos Campos
	•	Apresentar resposta do serviço de transferir (100%)

 4- Inserir Carga
	•	Remover saldo
	•	Titulo boleto de carga (100%)
	•	Botão copiar boleto
	•	Carga inserida = Boleto Gerado (100%)

 5- Cartao Virtual
	•	Trocar Botão por Switch de confirmação

 6- Tarifas
	•	Diminuir espaço entre os produtos
 7- Sair
	•	 notificação de confirmação (100%)

+++ app/src/main/java/itspay/br/com/model/FazerLoginPortador.java
@@ -10,8 +10,8 @@ public class FazerLoginPortador {
     private String deviceId;
     private long idInstituicao;
     private long idProcessadora;
-    private long latitude;
-    private long longitude;
+    private double latitude;
+    private double longitude;
     private String model;
     private long origemAcesso;
     private String plataformVersion;
@@ -23,7 +23,7 @@ public class FazerLoginPortador {
 
     public FazerLoginPortador(){}
 
-    public FazerLoginPortador(String architectureInfo, String cpf, String deviceId, long idInstituicao, long idProcessadora, long latitude, long longitude, String model, long origemAcesso, String plataformVersion, String platformName, String senha, long sistemaOperacional, String versaoConhecida, String versaoInstalada){
+    public FazerLoginPortador(String architectureInfo, String cpf, String deviceId, long idInstituicao, long idProcessadora, double latitude, double longitude, String model, long origemAcesso, String plataformVersion, String platformName, String senha, long sistemaOperacional, String versaoConhecida, String versaoInstalada){
         this.architectureInfo = architectureInfo;
         this.cpf = cpf;
         this.deviceId = deviceId;
@@ -61,11 +61,11 @@ public class FazerLoginPortador {
         return idProcessadora;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
@@ -121,11 +121,11 @@ public class FazerLoginPortador {
         this.idProcessadora = idProcessadora;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
